+++ wallet/core/src/wasm/tx/fees.rs
@@ -1,6 +1,7 @@
 use crate::imports::*;
 use crate::tx::fees::Fees;
 use kaspa_wallet_macros::declare_typescript_wasm_interface as declare;
+use wasm_bindgen::convert::TryFromJsValue;
 
 ///
 /// @see {@link IFees}, {@link IGeneratorSettingsObject}, {@link Generator}, {@link estimateTransactions}, {@link createTransactions}