+++ wasm/CHANGELOG.md
@@ -4,6 +4,7 @@ Updates since 0.3.14
 - Fix large RPC response deserialization errors in NodeJS caused by the default WebSocket frame size limit.
 - Fix event processing in UtxoContext
 - Renamed XPrivateKey to PrivateKeyGenerator and XPublicKey to PublicKeyGenerator
+- Renamed RpcClient.notify() to RpcClient.registerListener() / RpcClient.removeListener()
 - Simplify conversion between different key types (XPrv->Keypair, XPrv->XPub->Pubkey, etc)
 - Introduced `Beacon` class that provides connectivity to the community-operated public node infrastructure (backed by `kaspa-beacon` load balancer & node status monitor)
 - Created TypeScript type definitions across the entire SDK and refactored `RpcClient` class (as well as many other components) to use TypeScript interfaces